통계 기능 개선

+++ src/main/kotlin/com/group/libraryapp/controller/book/BookController.kt
@@ -32,7 +32,7 @@ class BookController(
     }
 
     @GetMapping("/book/loan")
-    fun countLoanedBook(): Int {
+    fun countLoanedBook(): Long {
         return bookService.countLoanedBook()
     }
