+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -142,12 +142,12 @@ class BookServiceTest @Autowired constructor(
 
     //then
     assertThat(results).hasSize(2)
-    assertCount(results, COMPUTER, 2)
-    assertCount(results, SCIENCE, 1)
+    assertCount(results, COMPUTER, 2L)
+    assertCount(results, SCIENCE, 1L)
   }
 
 }
 
-private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
   assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
 }
